blogProject/requirements.txt
dj-database-url==0.5.0
Django==1.11.12
django-widget-tweaks==1.4.2
Markdown==2.6.11
python-decouple==3.1
psycopg2==2.7.1
social-auth-app-django==2.1.0
django-allauth==0.35.0
djangorestframework
